[Refactor] 식단 운동 요약controller에 유저 유효성 검사 추가

+++ src/main/java/com/fcc/PureSync/context/exercise/controller/ExerciseController.java
@@ -3,6 +3,8 @@ package com.fcc.PureSync.context.exercise.controller;
 import com.fcc.PureSync.context.exercise.dto.ExerciseDto;
 import com.fcc.PureSync.context.exercise.dto.ExerciseResponseDto;
 import com.fcc.PureSync.common.ResultDto;
+import com.fcc.PureSync.exception.CustomException;
+import com.fcc.PureSync.exception.CustomExceptionCode;
 import com.fcc.PureSync.jwt.CustomUserDetails;
 import com.fcc.PureSync.context.exercise.service.ExerciseService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -22,12 +24,18 @@ public class ExerciseController {
 
     @GetMapping("/list")
     public ResultDto getAllExerciseList ( ExerciseDto exerciseTo, @AuthenticationPrincipal CustomUserDetails customUserDetails ) {
+        if (customUserDetails == null) {
+            throw new CustomException(CustomExceptionCode.INVALID_JWT);
+        }
         Long memSeq = customUserDetails.getMemSeq();
         return exerciseService.getExerciseAllList(exerciseTo, memSeq);
     }
 
     @PostMapping("/save")
     public  ResultDto exerciseInsert (@RequestBody ExerciseResponseDto exercise, @AuthenticationPrincipal CustomUserDetails customUserDetails ) {
+        if (customUserDetails == null) {
+            throw new CustomException(CustomExceptionCode.INVALID_JWT);
+        }
         Long memSeq = customUserDetails.getMemSeq();
         return exerciseService.insertExercise(exercise, memSeq);
     }
@@ -39,6 +47,9 @@ public class ExerciseController {
 
     @PostMapping("/delete")
     public  ResultDto exerciseDelete (@RequestBody ExerciseResponseDto exercise, @AuthenticationPrincipal CustomUserDetails customUserDetails ) {
+        if (customUserDetails == null) {
+            throw new CustomException(CustomExceptionCode.INVALID_JWT);
+        }
         Long memSeq = customUserDetails.getMemSeq();
         return exerciseService.deleteExercise(exercise, memSeq);
     }
